pyproject.toml
[tool.poetry]
name = "testgh"
version = "0.1.0"
description = ""
authors = ["tasi788 <tdc@tdccc.com.tw>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
PyNaCl = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
